Delete Exception handling

+++ src/main/java/com/ian/course/services/UserService.java
@@ -2,9 +2,12 @@ package com.ian.course.services;
 
 import com.ian.course.entities.User;
 import com.ian.course.repositories.UserRepository;
+import com.ian.course.services.exceptions.DatabaseException;
 import com.ian.course.services.exceptions.ResourceNotFoundException;
 import jakarta.annotation.Resource;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.security.PublicKey;
@@ -27,9 +30,18 @@ public class UserService {
     public User insert(User obj) {
         return repository.save(obj);
     }
-    public void delete(Long id){
-        repository.deleteById(id);
+    public void delete(Long id) {
+        try {
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
+
+
 
     public User update(Long id, User obj) {
         User entity = repository.getReferenceById(id);
